+++ RELEASE_NOTES.md
@@ -1,3 +1,11 @@
+## v0.4.0 (2022-07-28)
+
+- Completer
+  - A bubbling event is now dispatched when text entry is
+    finalized. See README and/or the Completer test page for more info
+- Debounce
+  - Fixed call to `l.log`
+
 ## v0.3.0 (2022-07-22)
 
 - Completer